requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
dill==0.3.0
enum34==1.1.6
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
mongoengine==0.18.2
multiprocess==0.70.8
pathlib2==2.3.3
pathos==0.2.4
pexpect==4.7.0
pickleshare==0.7.5
pika==1.0.1
pox==0.2.6
ppft==1.6.6.1
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.4.2
pymongo==3.8.0
PyYAML==5.1.1
requests==2.22.0
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
Top==1.0.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
